+++ lib/api.js
@@ -32,9 +32,10 @@ module.exports = settings => {
 
   app.use((req, res, next) => {
     if (res.response) {
-      const response = {
-        data: res.response
-      };
+      const response = {};
+      if (!req.query.onlymeta) {
+        response.data = res.response;
+      }
       response.meta = Object.assign({}, res.meta);
       if (req.establishment) {
         response.meta.establishment = {
