+++ lib/helpers/pils.js
@@ -1,0 +1,13 @@
+const moment = require('moment');
+
+function attachReviewDue(pil) {
+  return {
+    ...pil,
+    reviewDue: moment(pil.reviewDate).isBefore(moment().add(3, 'months')),
+    reviewOverdue: moment(pil.reviewDate).isBefore(moment())
+  };
+}
+
+module.exports = {
+  attachReviewDue
+};
