+++ lib/routers/establishment/index.js
@@ -117,7 +117,6 @@ router.put('/:establishment/grant',
   submit('grant')
 );
 
-router.use('/:establishment/pil-reviews', require('./pil-reviews'));
 router.use('/:establishment/role(s)?', require('./roles'));
 router.use('/:establishment/place(s)?', require('./places'));
 
@@ -126,7 +125,7 @@ router.use('/:establishment/profile(s)?', require('../profile'));
 router.use('/:establishment/project(s)?', require('./projects'));
 router.use('/:establishment/invite-user', require('./invite-user'));
 router.use('/:establishment/invitations', require('./invitations'));
-router.use('/:establishment/pil(s)?', require('../profile/pil'));
+router.use('/:establishment/pils', require('./pils'));
 router.use('/:establishment/billing', require('./billing'));
 
 module.exports = router;
