+++ lib/routers/profile/pil.js
@@ -1,10 +1,10 @@
+const { Router } = require('express');
 const { get } = require('lodash');
 const moment = require('moment');
-const { NotFoundError, BadRequestError } = require('../../errors');
+const { NotFoundError, BadRequestError, UnrecognisedActionError } = require('../../errors');
 const { fetchOpenTasks, permissions, validateSchema, whitelist, updateDataAndStatus } = require('../../middleware');
+const { attachReviewDue } = require('../../helpers/pils');
 const isUUID = require('uuid-validate');
-const { Router } = require('express');
-const { UnrecognisedActionError } = require('../../errors');
 
 const router = Router({ mergeParams: true });
 
@@ -69,29 +69,6 @@ const checkEstablishment = (req, res, next) => {
   next();
 };
 
-function getPils(req, res, next) {
-  const { PIL } = req.models;
-  const { search, sort, limit, offset } = req.query;
-
-  Promise.all([
-    PIL.count(req.establishment.id),
-    PIL.filter({
-      search,
-      sort,
-      limit,
-      offset,
-      establishmentId: req.establishment.id
-    })
-  ])
-    .then(([total, pils]) => {
-      res.meta.total = total;
-      res.meta.count = pils.total;
-      res.response = pils.results.map(attachReviewDue);
-      next();
-    })
-    .catch(next);
-}
-
 const attachEstablishmentDetails = (req, res, next) => {
   const establishmentId = req.pil.establishmentId;
   const { Establishment } = req.models;
@@ -109,14 +86,6 @@ const attachEstablishmentDetails = (req, res, next) => {
     .then(() => next())
     .catch(next);
 };
-
-function attachReviewDue(pil) {
-  return {
-    ...pil,
-    reviewDue: moment(pil.reviewDate).isBefore(moment().add(3, 'months')),
-    reviewOverdue: moment(pil.reviewDate).isBefore(moment())
-  };
-}
 
 router.param('pilId', (req, res, next, id) => {
   if (!isUUID(id)) {
@@ -215,11 +184,6 @@ router.delete('/:pilId',
   permissions('pil.delete'),
   checkEstablishment,
   submit('delete')
-);
-
-router.get('/',
-  permissions('pil.list'),
-  getPils
 );
 
 module.exports = router;
