+++ test/helpers/db.js
@@ -6,6 +6,7 @@ module.exports = settings => {
     init: (populate) => {
       const schema = Schema(settings);
       const tables = [
+        'AsruEstablishment',
         'ProjectVersion',
         'Project',
         'Permission',
